requirements.txt
Flask==2.3.3
boto3==1.34.63
psycopg2-binary==2.9.9